% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_functions.R
\name{get_kolada_data}
\alias{get_kolada_data}
\title{Get Data from Kolada API}
\usage{
get_kolada_data(kpi, municipality, year)
}
\arguments{
\item{kpi}{A string of the desired kpi statistic.}

\item{municipality}{a string of the desired municipality that user wants kpi statistic about.}

\item{year}{a string for the year the users want statistics about}
}
\value{
A list containing the API response data.
}
\description{
This function fetches data from the Kolada API.
}
