% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_municipality_id.R
\name{get_municipality}
\alias{get_municipality}
\title{Get ID for municipality}
\arguments{
\item{municipality_name}{String for the desired municipality to get id for}
}
\value{
The id for the input municipality
}
\description{
This function accepts a string for a municipality and returns the kolada ID for the given municipality
}
\examples{
\dontrun{
get_municipality("Stockholm")
}
}
