% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_municipality_id.R
\name{get_municipality_id}
\alias{get_municipality_id}
\title{Get ID for municipality}
\usage{
get_municipality_id(municipality_name)
}
\arguments{
\item{municipality_name}{String for the desired municipality to get id for}
}
\value{
The id for the input municipality
}
\description{
This function accepts a string for a municipality and returns the kolada ID for the given municipality
}
