\name{lab5-package}
\alias{lab5-package}
\alias{lab5}
\docType{package}
\title{
\packageTitle{lab5}
}
\description{
\packageDescription{lab5}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{lab5}

Maintainer: \packageMaintainer{lab5}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = lab5)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{lab5}}
%\section{Documentation Index}{\packageIndices{lab5}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
